Add Redis 2.8.19 to versions
Unlike the other formulas - we go ahead and alter the binary names so
that you can side-by-side with Redis 3.x (when that gets merged).

We don't alter the port number however so running both at the same time
won't work out of the box - but that's probably a good idea.